feat: Blueprint Runner MVP 1 (#889)
* chore: Adds tangle feature to blueprint-sdk

Enables the `tangle` feature for the `blueprint-sdk` dependency.
This addition is necessary to support upcoming functionalities that rely on tangle-related capabilities.

* refactor: key retrieval in squaring blueprint

Refactors the key retrieval process in the incredible squaring blueprint.

- Uses the keystore directly to fetch the SR25519 signer and secret, improving code clarity and efficiency.

* fix: Renames blueprint binary in run script

Updates the run script to reflect the renaming of the blueprint binary, ensuring the script correctly identifies and executes the intended binary.

* fix: tangle client creation with keystore

- Modifies `TangleClientContext` to use `with_keystore` for client creation, ensuring keystore integration.
- Simplifies keystore retrieval in `tangle/config.rs` by directly accessing the keystore from the environment.
- Removes redundant keystore path cloning.